requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.5.2
nltk==3.9.0
tensorflow-cpu==2.19.1
keras==2.19.0
